pyproject.toml
[project]
name = "tsunami-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Joscha Bisping",email = "joscha.l.bisping@tu-berlin.de"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "litellm (>=1.74.0,<2.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "langchain-anthropic (>=0.3.17,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "duckduckgo-search (==8.0.5)",
]

[tool.poetry]
packages = [{include = "tsunami_agent", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
